Add files via upload
The purpose of the project is to develop a simple source code editor for Microsoft Windows, that written in C#. Allows for quick and easy source code, and plain text editing. Allows recording the macros, and support more programming languages. It has many Principal Functions including MDI (Multiple Document Interface), Syntax Highlighting, Edit, Find, Replace as well as Zoom in and Zoom out. Now in the days of cross platform there are many languages that has to be supported by an editor, so we included support for many language like C#, Visual Basic, HTML, JavaScript, and SQL etc. There has been a lot of source editor. ProEdit designed to quickly and easily allow him to edit source code, to this end, a number of features. The ProEdit is "free software", free to modify and redistribute, and licensed under GPL. It is mainly based on Fast Coloured TextBox Editing component.